Cleanup go:build compiler directives
There's no need to maintain pre-go1.20 code, since we require go1.23+. Remove the dead code and cleanup `go:build` directives that are no longer needed.
