+++ pkg/backend/httpstate/client/marshal.go
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build go1.20
-
 package client
 
 import (
