+++ pkg/cmd/pulumi/stack/stack_output_fuzz_test.go
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build go1.18
-// +build go1.18
-
 package stack
 
 import (
